REFACTOR: convert component to view.
